refactor code to make numpy array export reuseable
